[#56][be][notification]refactor: 알림 타입 및 전송 상태 도메인 별 분리
