--- back/src/main/java/com/rouby/notification/domain/entity/SendStatus.java
@@ -1,7 +1,0 @@
-package com.rouby.notification.domain.entity;
-
-public enum SendStatus {
-  SEND,
-  FAILED
-  ;
-}
