+++ back/src/main/java/com/rouby/notification/domain/entity/emaillog/SendStatus.java
@@ -1,0 +1,7 @@
+package com.rouby.notification.domain.entity.emaillog;
+
+public enum SendStatus {
+  SEND,
+  FAILED
+  ;
+}
