+++ back/src/main/java/com/rouby/notification/domain/entity/notificationlog/SendStatus.java
@@ -1,0 +1,7 @@
+package com.rouby.notification.domain.entity.notificationlog;
+
+public enum SendStatus {
+  SEND,
+  FAILED
+  ;
+}
