+++ back/src/main/java/com/rouby/notificationtemplate/domain/entity/NotificationType.java
@@ -1,0 +1,9 @@
+package com.rouby.notificationtemplate.domain.entity;
+
+public enum NotificationType {
+
+  SCHEDULE,
+  ROUTINE,
+  BRIEFING,
+  ;
+}
